fixed database issue

+++ backend/src/services/vllm_client.py
@@ -81,11 +81,11 @@ class VLLMClient:
         # Extract text from output
         generated_text = outputs[0].outputs[0].text
         
-        # Store usage stats (approximate)
+        # Store usage stats (approximate) - MUST be integers for Pydantic validation
         self.last_usage = {
-            "prompt_tokens": len(prompt.split()) * 1.3,  # Rough estimate
-            "completion_tokens": len(generated_text.split()) * 1.3,  # Rough estimate
-            "total_tokens": (len(prompt) + len(generated_text)) * 1.3 / 4  # Rough estimate
+            "prompt_tokens": int(len(prompt.split()) * 1.3),  # Rough estimate
+            "completion_tokens": int(len(generated_text.split()) * 1.3),  # Rough estimate
+            "total_tokens": int((len(prompt) + len(generated_text)) * 1.3 / 4)  # Rough estimate
         }
         
         return generated_text
